Add supported Ruby versions - close #39
The specs were failing on newer versions of Ruby due to ActiveRecord 6. There was a hardcoded check for version 5, and this was causing a conditional spec to fallback to version 4. The spec has been updated to now test accordingly when the version is either below or above 5 (this is where the actual distinction lies).